% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.tree}
\alias{as.tree}
\title{Convert the result of a getTree call to a format compatible with tree}
\usage{
as.tree(gTree, rforest, max.depth = 3)
}
\arguments{
\item{gTree}{The results of a call to \code{getTree}}

\item{rforest}{The randomForest object}
}
\value{
An object of class \code{tree}, which has a \code{frame} and sufficient
    attributes to enable plotting
}
\description{
This function takes the results of a \code{randomForest::getTree} call and
converts the results to a form compatible with \code{tree}
}

