% Generated by roxygen2 (4.0.2): do not edit by hand
\name{int2bin}
\alias{int2bin}
\title{Convert integers to binary representation}
\usage{
int2bin(x, reverse = F)
}
\arguments{
\item{x}{integer to be converted}

\item{reverse}{Should the ordering be reversed}
}
\description{
Convert integers to binary representation
}

