% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict2.randomForest}
\alias{predict2.randomForest}
\title{Alternative version of predict.randomForest, from randomForest package}
\usage{
\method{predict2}{randomForest}(object, newdata, type = "response",
  norm.votes = TRUE, predict.all = FALSE, proximity = FALSE,
  nodes = FALSE, cutoff, ...)
}
\description{
This function adapts the predict.randomForest function to accomodate the situation
where factors can have NA as a legitimate level
}

