% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict2}
\alias{predict2}
\title{Alternative version of the predict function}
\usage{
predict2(object, ...)
}
\arguments{
\item{object}{Object from which predictions will be derived}
}
\description{
Alternative version of the predict function
}

