% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.reprtree.R
\name{text.tree}
\alias{text.tree}
\title{Annotate a Tree Plot}
\usage{
\method{text}{tree}(x, splits = TRUE, label = "yval", all = FALSE,
  pretty = NULL, digits = getOption("digits") - 3, adj = par("adj"),
  xpd = TRUE, uniform = T, ...)
}
\arguments{
\item{x}{an object of class "tree"}

\item{splits}{logical. If \code{TRUE} the splits are labelled}

\item{label}{The name of column in the \code{frame} component of \code{x}, to be used to label the nodes. Can be \code{NULL} to suppress node-labelling}

\item{all}{logical. By default, only the leaves are labelled, but if true interior nodes are also labelled}

\item{pretty}{the manipulation used for split labels infolving attributes. See Details.}

\item{digits}{significant digits for numerical labels}

\item{adj, xpd, ...}{graphical parameters such as \code{cex} and \code{font}}

\item{uniform}{logical. Is \code{plot(..., type='uniform')} used to create the tree plot?}
}
\description{
Add text to a tree plot. Modification of \code{tree:::text.tree} to add uniform type of tree.
}
\details{
If pretty = 0 then the level names of a factor split attributes are used unchanged. 
If pretty = NULL, the levels are presented by a, b, ... z, 0 ... 5. 
If pretty is a positive integer, abbreviate is applied to the labels with that value for its argument minlength.

If the lettering is vertical (par srt = 90) and adj is not supplied it is adjusted appropriately.
}
\seealso{
\code{\link{plot.tree}}
}
\author{
Abhijit Dasgupta, modifying original code by B.D. Ripley
}
