% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.reprtree}
\alias{plot.reprtree}
\title{Plotting representative trees}
\usage{
\method{plot}{reprtree}(reptree, all = F, index = ifelse(all, NULL, 1),
  depth = 0, main = NULL, adj = 0.5, ...)
}
\arguments{
\item{reptree}{An object of class \code{reprtree)}}

\item{all}{(logical) Do you want to create a panel of plots?}

\item{index}{The index of the reprtree object you want to plot}

\item{depth}{The maximum depth of the tree to be plotted. \code{depth=0} means the full tree.}

\item{main}{Title of plot (default is NULL)}

\item{adj}{}

\item{...}{additional arguments to pass to text.tree. In particular, suppress node labels using \code{label=NULL}}
}
\description{
This function creates either a single plot or a panel of plots for
visualizing representative trees from an ensemble
}
\section{Details}{

This plot function takes a \code{reprtree} object, and then plots a
single representative tree or a sequence of representative trees (using \code{all=T}).

If only one tree needs to be visualized, the index of the reprtree object to
be visualized can be provided.
}

